refactor(http-server): use mods, reduce code duplication

better test
